modify: 컴포넌트 내 타입을 types 파일로 분리
- 컴포넌트에 있던 타입 정의를 types 폴더로 이동
